+++ Projects/ProjectARVRPro/Process/Chessboard/ChessboardDynamicProcess.cs
@@ -16,7 +16,7 @@ namespace ProjectARVRPro.Process.Chessboard
 {
     public class ChessboardDynamicProcess : ProcessBase<ChessboardDynamicProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;