+++ Projects/ProjectARVRPro/Process/Distortion/DistortionDynamicProcess.cs
@@ -12,7 +12,7 @@ namespace ProjectARVRPro.Process.Distortion
 {
     public class DistortionDynamicProcess : ProcessBase<DistortionDynamicProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;