+++ Projects/ProjectARVRPro/Process/Distortion/DistortionProcess.cs
@@ -11,7 +11,7 @@ namespace ProjectARVRPro.Process.Distortion
 {
     public class DistortionProcess : ProcessBase<DistortionProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;