+++ Projects/ProjectARVRPro/Process/MTF/MTFH/MTFHProcess.cs
@@ -19,7 +19,7 @@ namespace ProjectARVRPro.Process.MTF.MTFH
     /// </summary>
     public class MTFHProcess : ProcessBase<MTFHProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;