+++ Projects/ProjectARVRPro/Process/MTF/MTFHV/MTFHVProcess.cs
@@ -15,7 +15,7 @@ namespace ProjectARVRPro.Process.MTF.MTFHV
 {
     public class MTFHVProcess : ProcessBase<MTFHVProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;