+++ Projects/ProjectARVRPro/Process/MTF/MTFV/MTFVProcess.cs
@@ -19,7 +19,7 @@ namespace ProjectARVRPro.Process.MTF.MTFV
     /// </summary>
     public class MTFVProcess : ProcessBase<MTFVProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;