+++ Projects/ProjectARVRPro/Process/OpticCenter/OpticCenterProcess.cs
@@ -10,7 +10,7 @@ namespace ProjectARVRPro.Process.OpticCenter
 {
     public class OpticCenterProcess : ProcessBase<OpticCenterProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;