+++ Projects/ProjectARVRPro/Process/RGB/Red/RedProcess.cs
@@ -18,7 +18,7 @@ namespace ProjectARVRPro.Process.RGB.Red
 
     public class RedProcess : ProcessBase<RedProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;