+++ Projects/ProjectARVRPro/Process/W255/White255Process.cs
@@ -18,7 +18,7 @@ namespace ProjectARVRPro.Process.W255
 {
     public class White255Process : ProcessBase<W255ProcessConfig>
     {
-        public override bool Execute(IProcessExecutionContext ctx)
+        public override async Task<bool> Execute(IProcessExecutionContext ctx)
         {
             if (ctx?.Batch == null || ctx.Result == null) return false;
             var log = ctx.Log;